class extends Store
