+++ src/alt/index.js
@@ -3,9 +3,12 @@ import { Dispatcher } from 'flux'
 
 import * as StateFunctions from './utils/StateFunctions'
 import * as fn from '../utils/functions'
-import * as store from './store'
 import * as utils from './utils/AltUtils'
 import makeAction from './actions'
+
+import Store from './store'
+
+export { Store }
 
 class Alt {
   constructor(config = {}) {
@@ -25,44 +28,6 @@ class Alt {
 
   dispatch(action, data, details) {
     this.batchingFunction(() => this.dispatcher.dispatch({ action, data, details }))
-  }
-
-  createUnsavedStore(StoreModel, ...args) {
-    const key = StoreModel.displayName || ''
-    store.createStoreConfig(this.config, StoreModel)
-    const Store = store.transformStore(this.storeTransforms, StoreModel)
-
-    return fn.isFunction(Store)
-      ? store.createStoreFromClass(this, Store, key, ...args)
-      : store.createStoreFromObject(this, Store, key)
-  }
-
-  createStore(StoreModel, iden, ...args) {
-    let key = iden || StoreModel.displayName || StoreModel.name || ''
-    store.createStoreConfig(this.config, StoreModel)
-    const Store = store.transformStore(this.storeTransforms, StoreModel)
-
-    if (this.stores[key] || !key) {
-      if (this.stores[key]) {
-        utils.warn(
-          `A store named ${key} already exists, double check your store ` +
-          `names or pass in your own custom identifier for each store`
-        )
-      } else {
-        utils.warn('Store name was not specified')
-      }
-
-      key = utils.uid(this.stores, key)
-    }
-
-    const storeInstance = fn.isFunction(Store)
-      ? store.createStoreFromClass(this, Store, key, ...args)
-      : store.createStoreFromObject(this, Store, key)
-
-    this.stores[key] = storeInstance
-    StateFunctions.saveInitialSnapshot(this, key)
-
-    return storeInstance
   }
 
   generateActions(...actionNames) {
